+++ src/InputReader.h
@@ -149,7 +149,7 @@ struct InputZeroMQThreadData
 {
     ThreadsafeQueue<std::shared_ptr<std::vector<uint8_t> > > *in_messages;
     std::string uri;
-    unsigned max_queued_frames;
+    size_t max_queued_frames;
 };
 
 class InputZeroMQWorker
@@ -194,7 +194,7 @@ class InputZeroMQReader : public InputReader
             worker_.Stop();
         }
 
-        int Open(const std::string& uri, unsigned max_queued_frames);
+        int Open(const std::string& uri, size_t max_queued_frames);
 
         int GetNextFrame(void* buffer);
